% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_gaussian_process.R
\name{logistic_gaussian_process}
\alias{logistic_gaussian_process}
\title{Fit a logistic Gaussian process in R for a binary outcome}
\usage{
logistic_gaussian_process(
  formula,
  data,
  sparse = T,
  n_points = 10,
  n_particles = 10,
  n_iter = 2000
)
}
\arguments{
\item{data}{the dataframe from which to source regression data.}

\item{sparse}{a Boolean indicating whether to use the projected process
approximation to speed up training. Recommended for N>100.}

\item{n_points}{The number of inducing points used for the projected process
approximation. Defaults to 5. If set too large, may cause numerical
stability issues when calculating the gradient of the marginal likelihood.}

\item{n_particles}{number of particles to be used in the stein variational gradient descent simulation}

\item{form}{the regression formula for the Gaussian process. The left-hand-side
gives determines outcome you are fitting the GP to, and the right-hand side gives the kernel specification.}
}
\description{
Fit a logistic Gaussian process in R for a binary outcome
}
