% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gp.R
\name{plot.gaussian_process}
\alias{plot.gaussian_process}
\title{Plot Gaussian Process Fit}
\usage{
\method{plot}{gaussian_process}(x, col = 1, add_points = T, ...)
}
\arguments{
\item{x}{Gaussian Process object}

\item{col}{column name for the predictor you want to plot. All other
predictors are set to their medians}

\item{add_points}{whether to plot datapoints in the background}

\item{...}{additional arguments (ignored)}
}
\description{
Plot Gaussian Process Fit
}
