% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_process_methods.R
\name{predict.gaussian_process}
\alias{predict.gaussian_process}
\title{Predict Function Values at a Test Point}
\usage{
\method{predict}{gaussian_process}(object, newdata = NULL, kernel_str = NA, ...)
}
\arguments{
\item{newdata}{a new data set to generate predictions for. Must have all of
the predictor columns of the original data set or an error will be returned.}

\item{kernel_str}{the label of the kernel you would like to extract
predictions for. This can be useful if you are (for example) attempting to
isolate a specific signal from a data set such as a periodic component from a
time series.  Use the print method of the Gaussian process to determine the
labels for the individual kernels.}
}
\description{
Predict Function Values at a Test Point
}
