% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_process_methods.R
\name{print.gaussian_process}
\alias{print.gaussian_process}
\title{Print information about a gaussian process}
\usage{
\method{print}{gaussian_process}(x, ...)
}
\description{
Print information about a gaussian process
}
