% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_process_methods.R
\name{residuals.gaussian_process}
\alias{residuals.gaussian_process}
\title{Get Residuals for Gaussian Process}
\usage{
\method{residuals}{gaussian_process}(object, ...)
}
\arguments{
\item{object}{An object for which residuals are desired.}

\item{...}{additional arguments (ignored)}
}
\description{
Get Residuals for Gaussian Process
}
